Move script types to /models
