+++ src/models/script/script.ts
